requirements.txt
pika==0.9.8
nose
requests
dingus
microhackaton-service-discovery-py
